Pass Optimal to switcher's init
Since Optimal also reads from the same address stream, and it needs to show the page faults for the same set of pages as the other algorithms, it is necessary to perform the same sleep operation on Optimal as well after the defined number of page accesses.

But since Optimal is not a candidate for switching, it is not passed as a part of the other_algorithm[] list. Hence, the resetting of its local page fault count and number of pages accessed is done outside the loop operating on other_algorithm[]